prevent go routine managing reconnects to leak on fsm stop
