+++ protocols/bgp/server/server.go
@@ -2,13 +2,11 @@ package server
 
 import (
 	"fmt"
-	"io"
 	"net"
 	"strings"
 	"sync"
 
 	"github.com/bio-routing/bio-rd/config"
-	"github.com/bio-routing/bio-rd/protocols/bgp/packet"
 	"github.com/bio-routing/bio-rd/routingtable/locRIB"
 	log "github.com/sirupsen/logrus"
 )
@@ -124,21 +122,4 @@ func (b *bgpServer) AddPeer(c config.Peer, rib *locRIB.LocRIB) error {
 	peer.Start()
 
 	return nil
-}
-
-func recvMsg(c net.Conn) (msg []byte, err error) {
-	buffer := make([]byte, packet.MaxLen)
-	_, err = io.ReadFull(c, buffer[0:packet.MinLen])
-	if err != nil {
-		return nil, fmt.Errorf("Read failed: %v", err)
-	}
-
-	l := int(buffer[16])*256 + int(buffer[17])
-	toRead := l
-	_, err = io.ReadFull(c, buffer[packet.MinLen:toRead])
-	if err != nil {
-		return nil, fmt.Errorf("Read failed: %v", err)
-	}
-
-	return buffer, nil
 }
